Port SM103 kernels and validation to preparation sessions
Retain SM103 support on master 213fc1b204b306bdbaa7d40d2a27529658128bf7
through declarative plans, typed tuning contracts, offline compile jobs,
private materialized state and session-owned program lifetimes.

Features retained or exposed through preparation:
- Native CuTe tcgen05/TMEM NVFP4 and Trellis MoE, including uniform,
  coupled, mixed-rate, grouped atoms and canonical BtxSource/BtxWeights.
- Dense NVFP4/MXFP4/MXFP8, MXFP6/W6A8 and tensor/block FP8, bounded FP8
  workspace, block-FP8 linear, WO and vocabulary projection.
- GLM sparse MLA, compressed MLA and DSA indexing; KDA/GDN, mHC,
  HyperConnection, three MTP contracts and Engram storage ownership.
- Metadata-only SM103 compiler coverage, native PTX/cubin/SASS/resource
  evidence, deferred hardware qualification and portable serving tests.

Required fixes:
- Carry architecture lowering and exact program identities through planning,
  compilation, materialization and binding. Use capacity specializations
  with runtime live counts and validate architecture admission centrally.
- Give mixed-rate Trellis and DSA cache entries explicit program carriers.
  Evict DSA discovery placeholders before in-process compilation and require
  one native export for every declared CuTe program in the offline audit.
- Preserve graph-visible output/scratch mutation through typed opaque ops
  for attention, block-FP8, FP8 workspace and concatenated/FP8 MTP paths.
- Keep allocation, quantization and bias on the requested stream; handle
  empty requests without zero-sized grids; retain caller-owned storage.
- Preserve FP16 packed MXFP8 query/output types, use FP32 split partials
  with typed reduction and restrict BF16 atomics to BF16 outputs.
- Include the attention sink in sparse MLA LSE; retain Trellis broadcast
  metadata, grouped rates, global transform coordinates and input halves.
- Preserve Int64 pool/vocabulary offsets and verify manifest-bound cache
  objects and launch metadata before loading temporary ELF copies.

Compatibility: integrations use PreparationSession and canonical weight and
execution plans. The removed policy/prewarm interfaces are not restored.
Engram disk transactions follow upstream synchronous semantics; compatibility
prefetch methods do not supply asynchronous prefetch. The companion vLLM
branch requires its own preparation-API port before full-model validation.

Validation includes 82 preparation declarations with all 221 declared CuTe
programs exported, 1,163 native callables with 1,165 CUDA entry points,
877 host passes and 1,093 portable GPU passes. Exact
source/artifact identities and all failures/skips are recorded in
docs/sm103-preparation-validation.json; the feature/fix map and qualification
runbook describe the public contracts and remaining limits. Three preparation
test failures reproduce on pristine master; no physical SM103 execution,
B300 performance or full-model accuracy qualification is claimed.

+++ b12x/gemm/blockscaled/_ops.py
@@ -25,7 +25,7 @@ def _execute(source, values, scales, global_scale, activation_scale, workspace, 
 
 
 def _metadata(source, values):
-    return source.new_empty((*source.shape[:-1], values.shape[0]), dtype=torch.bfloat16)
+    return source.new_empty((*source.shape[:-1], values.shape[0]), dtype=source.dtype)
 
 
 @torch.library.custom_op("b12x::blockscaled_bf16", mutates_args=())
